TransactionalMasterRequestListener: move the extended logic directly to the framework package
