+++ packages/framework/src/Component/HttpFoundation/TransactionalMasterRequestListener.php
@@ -5,12 +5,18 @@ declare(strict_types=1);
 namespace Shopsys\FrameworkBundle\Component\HttpFoundation;
 
 use Doctrine\ORM\EntityManagerInterface;
+use GraphQL\Error\SyntaxError;
+use GraphQL\Language\AST\OperationDefinitionNode;
+use GraphQL\Language\Parser;
 use Symfony\Component\HttpKernel\Event\ExceptionEvent;
 use Symfony\Component\HttpKernel\Event\RequestEvent;
 use Symfony\Component\HttpKernel\Event\ResponseEvent;
 
 class TransactionalMasterRequestListener
 {
+    protected const GRAPHQL_ENDPOINT_ROUTE = 'overblog_graphql_endpoint';
+    protected const QUERY_TYPE = 'query';
+
     protected bool $inTransaction = false;
 
     protected static bool $isManuallyRollbacked = false;
@@ -33,7 +39,7 @@ class TransactionalMasterRequestListener
      */
     public function onKernelRequest(RequestEvent $event): void
     {
-        if ($event->isMainRequest() && !$this->inTransaction) {
+        if ($event->isMainRequest() && !$this->inTransaction && !$this->isRequestGraphQlQuery($event)) {
             $this->em->beginTransaction();
             $this->inTransaction = true;
         }
@@ -59,5 +65,44 @@ class TransactionalMasterRequestListener
             $this->em->rollback();
             $this->inTransaction = false;
         }
+    }
+
+    /**
+     * @param \Symfony\Component\HttpKernel\Event\RequestEvent $requestEvent
+     * @return bool
+     */
+    protected function isRequestGraphQlQuery(RequestEvent $requestEvent): bool
+    {
+        if ($requestEvent->getRequest()->attributes->get('_route') !== static::GRAPHQL_ENDPOINT_ROUTE) {
+            return false;
+        }
+
+        $requestContent = $requestEvent->getRequest()->getContent();
+
+        if ($requestContent === null || $requestContent === '') {
+            return false;
+        }
+
+        $source = json_decode($requestContent, true);
+
+        if (!array_key_exists(static::QUERY_TYPE, $source)) {
+            return false;
+        }
+
+        $queryString = $source[static::QUERY_TYPE];
+
+        try {
+            $parsed = Parser::parse($queryString);
+
+            foreach ($parsed->definitions as $definition) {
+                if ($definition instanceof OperationDefinitionNode) {
+                    return $definition->operation === static::QUERY_TYPE;
+                }
+            }
+        } catch (SyntaxError) {
+            return false;
+        }
+
+        return false;
     }
 }
