+++ packages/frontend-api/src/Resources/config/services.yaml
@@ -19,13 +19,6 @@ services:
     Lcobucci\JWT\Configuration:
         factory: ['@Shopsys\FrontendApiBundle\Model\Token\JwtConfigurationFactory', create]
 
-    Shopsys\FrameworkBundle\Component\HttpFoundation\TransactionalMasterRequestListener:
-        class: Shopsys\FrontendApiBundle\Component\HttpFoundation\TransactionalMasterRequestListener
-        tags:
-            - { name: kernel.event_listener, event: kernel.request, method: onKernelRequest }
-            - { name: kernel.event_listener, event: kernel.response, method: onKernelResponse }
-            - { name: kernel.event_listener, event: kernel.exception, method: onKernelException }
-
     Shopsys\FrontendApiBundle\Component\Domain\EnabledOnDomainChecker:
         arguments:
             $enabledDomainIds: '%shopsys.frontend_api.domains%'
